+++ hapi-fhir-base/src/main/java/ca/uhn/fhir/util/BundleUtil.java
@@ -25,7 +25,6 @@ import ca.uhn.fhir.context.BaseRuntimeElementDefinition;
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.RuntimeResourceDefinition;
 import ca.uhn.fhir.i18n.Msg;
-import ca.uhn.fhir.model.api.StorageResponseCodeEnum;
 import ca.uhn.fhir.model.primitive.IdDt;
 import ca.uhn.fhir.model.valueset.BundleTypeEnum;
 import ca.uhn.fhir.rest.api.PatchTypeEnum;
@@ -47,7 +46,6 @@ import org.hl7.fhir.instance.model.api.IBaseBinary;
 import org.hl7.fhir.instance.model.api.IBaseBundle;
 import org.hl7.fhir.instance.model.api.IBaseReference;
 import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.hl7.fhir.instance.model.api.IIdType;
 import org.hl7.fhir.instance.model.api.IPrimitiveType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -877,143 +875,6 @@ public class BundleUtil {
 						.equals(theResource.getIdElement().getIdPart()))
 				.findFirst()
 				.orElse(null);
-	}
-
-	/**
-	 * This method accepts a {@literal Bundle} which was returned by a call to HAPI FHIR's
-	 * transaction/batch processor. HAPI FHIR has specific codes and extensions it will
-	 * always put into the OperationOutcomes returned by the transaction processor, so
-	 * this method parses these and returns a more machine-processable interpretation.
-	 * <p>
-	 * This method should only be called for Bundles returned by HAPI FHIR's transaction
-	 * processor, results will have no meaning for any other input.
-	 * </p>
-	 *
-	 * @since 8.2.0
-	 */
-	public static TransactionResponse toTransactionResponse(
-			FhirContext theContext, IBaseBundle theTransactionResponseBundle) {
-		FhirTerser terser = theContext.newTerser();
-		List<StorageOutcome> storageOutcomes = new ArrayList<>();
-
-		List<IBase> entries = terser.getValues(theTransactionResponseBundle, "entry");
-		for (IBase entry : entries) {
-
-			IBase response = terser.getSingleValueOrNull(entry, "response", IBase.class);
-			if (response != null) {
-
-				String statusString = terser.getSinglePrimitiveValueOrNull(response, "status");
-				int statusSpaceIdx = statusString.indexOf(' ');
-				int statusCode = Integer.parseInt(statusString.substring(0, statusSpaceIdx));
-
-				List<IBase> issues = terser.getValues(response, "outcome.issue");
-				IIdType groupSourceId = null;
-				for (int issueIndex = 0; issueIndex < issues.size(); issueIndex++) {
-					IBase issue = issues.get(issueIndex);
-					IIdType sourceId = null;
-
-					String outcomeSystem = terser.getSinglePrimitiveValueOrNull(issue, "details.coding.system");
-					StorageResponseCodeEnum responseCode = null;
-					if (StorageResponseCodeEnum.SYSTEM.equals(outcomeSystem)) {
-						String outcomeCode = terser.getSinglePrimitiveValueOrNull(issue, "details.coding.code");
-						responseCode = StorageResponseCodeEnum.valueOf(outcomeCode);
-					}
-
-					String errorMessage = null;
-					String issueCode = terser.getSinglePrimitiveValueOrNull(issue, "code");
-					if ("exception".equals(issueCode)) {
-						errorMessage = terser.getSinglePrimitiveValueOrNull(issue, "diagnostics");
-					}
-
-					IIdType targetId = null;
-					if (responseCode == StorageResponseCodeEnum.AUTOMATICALLY_CREATED_PLACEHOLDER_RESOURCE) {
-						/*
-						TODO: uncomment this when branch ja_20250217_tx_log_provenance merges
-						targetId = ((IBaseHasExtensions) issue)
-							.getExtension().stream()
-							.filter(t -> HapiExtensions.EXTENSION_PLACEHOLDER_ID.equals(t.getUrl()))
-							.findFirst()
-							.map(t -> (IIdType) t.getValue())
-							.orElse(null);
-						*/
-						sourceId = groupSourceId;
-					} else {
-						targetId = theContext
-								.getVersion()
-								.newIdType(terser.getSinglePrimitiveValueOrNull(entry, "response.location"));
-						if (issueIndex == 0) {
-							groupSourceId = targetId;
-						}
-					}
-
-					StorageOutcome outcome =
-							new StorageOutcome(statusCode, responseCode, targetId, sourceId, errorMessage);
-					storageOutcomes.add(outcome);
-				}
-			}
-		}
-
-		return new TransactionResponse(storageOutcomes);
-	}
-
-	/**
-	 * @see #toTransactionResponse(FhirContext, IBaseBundle)
-	 */
-	public static class TransactionResponse {
-
-		private final List<StorageOutcome> myStorageOutcomes;
-
-		public TransactionResponse(List<StorageOutcome> theStorageOutcomes) {
-			myStorageOutcomes = theStorageOutcomes;
-		}
-
-		public List<StorageOutcome> getStorageOutcomes() {
-			return myStorageOutcomes;
-		}
-	}
-
-	/**
-	 * @see #toTransactionResponse(FhirContext, IBaseBundle)
-	 */
-	public static class StorageOutcome {
-		private final StorageResponseCodeEnum myStorageResponseCode;
-		private final IIdType myTargetId;
-		private final IIdType mySourceId;
-		private final int myStatusCode;
-		private final String myErrorMessage;
-
-		public StorageOutcome(
-				int theStatusCode,
-				StorageResponseCodeEnum theStorageResponseCode,
-				IIdType theTargetId,
-				IIdType theSourceId,
-				String theErrorMessage) {
-			myStatusCode = theStatusCode;
-			myStorageResponseCode = theStorageResponseCode;
-			myTargetId = theTargetId;
-			mySourceId = theSourceId;
-			myErrorMessage = theErrorMessage;
-		}
-
-		public String getErrorMessage() {
-			return myErrorMessage;
-		}
-
-		public int getStatusCode() {
-			return myStatusCode;
-		}
-
-		public StorageResponseCodeEnum getStorageResponseCode() {
-			return myStorageResponseCode;
-		}
-
-		public IIdType getTargetId() {
-			return myTargetId;
-		}
-
-		public IIdType getSourceId() {
-			return mySourceId;
-		}
 	}
 
 	private static class SortLegality {
