+++ hapi-fhir-storage/src/main/java/ca/uhn/fhir/jpa/dao/BaseTransactionProcessor.java
@@ -225,67 +225,7 @@ public abstract class BaseTransactionProcessor {
 	public <BUNDLE extends IBaseBundle> BUNDLE transaction(
 			RequestDetails theRequestDetails, BUNDLE theRequest, boolean theNestedMode) {
 		String actionName = "Transaction";
-
-		TransactionSemanticsHeader transactionSemantics = TransactionSemanticsHeader.DEFAULT;
-		if (theRequestDetails != null) {
-			String transactionSemanticsString = theRequestDetails.getHeader("X-Transaction-Semantics");
-			if (transactionSemanticsString != null) {
-				transactionSemantics = TransactionSemanticsHeader.parse(transactionSemanticsString);
-			}
-		}
-
-		int totalAttempts = defaultIfNull(transactionSemantics.getRetryCount(), 0) + 1;
-		boolean switchedToBatch = false;
-
-		IBaseBundle response;
-		for (int i = 1; ; i++) {
-			try {
-				response = processTransactionAsSubRequest(theRequestDetails, theRequest, actionName, theNestedMode);
-				break;
-			} catch (BaseServerResponseException e) {
-				if (i >= totalAttempts || theNestedMode) {
-					throw e;
-				}
-
-				int minRetryDelay = defaultIfNull(transactionSemantics.getMinRetryDelay(), 0);
-				int maxRetryDelay = defaultIfNull(transactionSemantics.getMaxRetryDelay(), 0);
-				maxRetryDelay = Math.max(maxRetryDelay, minRetryDelay);
-
-				// Don't let the user request a crazy delay, this could be used
-				// as a DOS attack
-				maxRetryDelay = Math.min(maxRetryDelay, 10000);
-
-				long delay = RandomUtils.insecure().randomLong(minRetryDelay, maxRetryDelay);
-				String sleepMessage = "";
-				if (delay > 0) {
-					sleepMessage = "Sleeping for " + delay + " ms. ";
-				}
-
-				String switchedToBatchMessage = "";
-				if (switchedToBatch) {
-					switchedToBatchMessage = " (as batch)";
-				}
-
-				String switchingToBatchMessage = "";
-				if (i + 1 == totalAttempts && transactionSemantics.isFinalRetryAsBatch()) {
-					switchingToBatchMessage = "Performing final retry using batch semantics. ";
-					switchedToBatch = true;
-					BundleUtil.setBundleType(myContext, theRequest, "batch");
-				}
-
-				ourLog.warn(
-						"Transaction processing attempt{} {}/{} failed. {}{}",
-						switchedToBatchMessage,
-						i,
-						totalAttempts,
-						sleepMessage,
-						switchingToBatchMessage);
-
-				if (delay > 0) {
-					ThreadUtils.sleepQuietly(Duration.ofMillis(delay));
-				}
-			}
-		}
+		IBaseBundle response = processTransactionAsSubRequest(theRequestDetails, theRequest, actionName, theNestedMode);
 
 		List<IBase> entries = myVersionAdapter.getEntries(response);
 		for (int i = 0; i < entries.size(); i++) {
@@ -451,6 +391,77 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	private IBaseBundle batch(final RequestDetails theRequestDetails, IBaseBundle theRequest, boolean theNestedMode) {
+		TransactionSemanticsHeader transactionSemantics = TransactionSemanticsHeader.DEFAULT;
+		if (theRequestDetails != null) {
+			String transactionSemanticsString = theRequestDetails.getHeader("X-Transaction-Semantics");
+			if (transactionSemanticsString != null) {
+				transactionSemantics = TransactionSemanticsHeader.parse(transactionSemanticsString);
+			}
+		}
+
+		int totalAttempts = defaultIfNull(transactionSemantics.getRetryCount(), 0) + 1;
+		boolean switchedToBatch = false;
+
+		IBaseBundle response;
+		for (int i = 1; ; i++) {
+			try {
+				if (i < totalAttempts && transactionSemantics.isTryBatchAsTransactionFirst()) {
+					BundleUtil.setBundleType(myContext, theRequest, "transaction");
+					response = processTransaction(theRequestDetails, theRequest, "Transaction", theNestedMode);
+				} else {
+					BundleUtil.setBundleType(myContext, theRequest, "batch");
+					response = processBatch(theRequestDetails, theRequest, theNestedMode);
+				}
+				break;
+			} catch (BaseServerResponseException e) {
+				if (i >= totalAttempts || theNestedMode) {
+					throw e;
+				}
+
+				int minRetryDelay = defaultIfNull(transactionSemantics.getMinRetryDelay(), 0);
+				int maxRetryDelay = defaultIfNull(transactionSemantics.getMaxRetryDelay(), 0);
+				maxRetryDelay = Math.max(maxRetryDelay, minRetryDelay);
+
+				// Don't let the user request a crazy delay, this could be used
+				// as a DOS attack
+				maxRetryDelay = Math.min(maxRetryDelay, 10000);
+
+				long delay = RandomUtils.insecure().randomLong(minRetryDelay, maxRetryDelay);
+				String sleepMessage = "";
+				if (delay > 0) {
+					sleepMessage = "Sleeping for " + delay + " ms. ";
+				}
+
+				String switchedToBatchMessage = "";
+				if (switchedToBatch) {
+					switchedToBatchMessage = " (as batch)";
+				}
+
+				String switchingToBatchMessage = "";
+				if (i + 1 == totalAttempts && transactionSemantics.isTryBatchAsTransactionFirst()) {
+					switchingToBatchMessage = "Performing final retry using batch semantics. ";
+					switchedToBatch = true;
+					BundleUtil.setBundleType(myContext, theRequest, "batch");
+				}
+
+				ourLog.warn(
+					"Transaction processing attempt{} {}/{} failed. {}{}",
+					switchedToBatchMessage,
+					i,
+					totalAttempts,
+					sleepMessage,
+					switchingToBatchMessage);
+
+				if (delay > 0) {
+					ThreadUtils.sleepQuietly(Duration.ofMillis(delay));
+				}
+			}
+		}
+
+		return response;
+	}
+
+	private IBaseBundle processBatch(RequestDetails theRequestDetails, IBaseBundle theRequest, boolean theNestedMode) {
 		ourLog.info(
 				"Beginning batch with {} resources",
 				myVersionAdapter.getEntries(theRequest).size());
@@ -730,14 +741,14 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * All of the write operations in the transaction (PUT, POST, etc.. basically anything
-	 * except GET) are performed in their own database transaction before we do the reads.
-	 * We do this because the reads (specifically the searches) often spawn their own
-	 * secondary database transaction and if we allow that within the primary
-	 * database transaction we can end up with deadlocks if the server is under
-	 * heavy load with lots of concurrent transactions using all available
-	 * database connections.
-	 */
+     * All of the write operations in the transaction (PUT, POST, etc.. basically anything
+     * except GET) are performed in their own database transaction before we do the reads.
+     * We do this because the reads (specifically the searches) often spawn their own
+     * secondary database transaction and if we allow that within the primary
+     * database transaction we can end up with deadlocks if the server is under
+     * heavy load with lots of concurrent transactions using all available
+     * database connections.
+     */
 	@SuppressWarnings("unchecked")
 	private void prepareThenExecuteTransactionWriteOperations(
 			RequestDetails theRequestDetails,
@@ -810,9 +821,9 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * This method looks at the FHIR actions being performed in a List of bundle entries,
-	 * and determines the associated request partitions.
-	 */
+     * This method looks at the FHIR actions being performed in a List of bundle entries,
+     * and determines the associated request partitions.
+     */
 	@Nullable
 	protected RequestPartitionId determineRequestPartitionIdForWriteEntries(
 			RequestDetails theRequestDetails, List<IBase> theEntries) {
@@ -948,13 +959,13 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * This method is called for nested bundles (e.g. if we received a transaction with an entry that
-	 * was a GET search, this method is called on the bundle for the search result, that will be placed in the
-	 * outer bundle). This method applies the _summary and _content parameters to the output of
-	 * that bundle.
-	 * <p>
-	 * TODO: This isn't the most efficient way of doing this.. hopefully we can come up with something better in the future.
-	 */
+     * This method is called for nested bundles (e.g. if we received a transaction with an entry that
+     * was a GET search, this method is called on the bundle for the search result, that will be placed in the
+     * outer bundle). This method applies the _summary and _content parameters to the output of
+     * that bundle.
+     * <p>
+     * TODO: This isn't the most efficient way of doing this.. hopefully we can come up with something better in the future.
+     */
 	private IBaseResource filterNestedBundle(RequestDetails theRequestDetails, IBaseResource theResource) {
 		IParser p = myContext.newJsonParser();
 		RestfulServerUtils.configureResponseParser(theRequestDetails, p);
@@ -971,8 +982,8 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Searches for duplicate conditional creates and consolidates them.
-	 */
+     * Searches for duplicate conditional creates and consolidates them.
+     */
 	@SuppressWarnings("unchecked")
 	private void consolidateDuplicateConditionals(
 			RequestDetails theRequestDetails, String theActionName, List<IBase> theEntries) {
@@ -1063,9 +1074,9 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Iterates over all entries, and if it finds any which have references which match the fullUrl of the entry that was consolidated out
-	 * replace them with our new consolidated UUID
-	 */
+     * Iterates over all entries, and if it finds any which have references which match the fullUrl of the entry that was consolidated out
+     * replace them with our new consolidated UUID
+     */
 	private void replaceReferencesInEntriesWithConsolidatedUUID(
 			List<IBase> theEntries, String theEntryFullUrl, String existingUuid) {
 		for (IBase nextEntry : theEntries) {
@@ -1091,14 +1102,14 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Retrieves the next resource id (IIdType) from the base resource and next request entry.
-	 *
-	 * @param theBaseResource - base resource
-	 * @param theNextReqEntry - next request entry
-	 * @param theAllIds       - set of all IIdType values
-	 * @param theVerb
-	 * @return
-	 */
+     * Retrieves the next resource id (IIdType) from the base resource and next request entry.
+     *
+     * @param theBaseResource - base resource
+     * @param theNextReqEntry - next request entry
+     * @param theAllIds       - set of all IIdType values
+     * @param theVerb
+     * @return
+     */
 	private IIdType getNextResourceIdFromBaseResource(
 			IBaseResource theBaseResource, IBase theNextReqEntry, Set<IIdType> theAllIds, String theVerb) {
 		IIdType nextResourceId = null;
@@ -1165,8 +1176,8 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * After pre-hooks have been called
-	 */
+     * After pre-hooks have been called
+     */
 	@SuppressWarnings({"unchecked", "rawtypes"})
 	protected EntriesToProcessMap doTransactionWriteOperations(
 			final RequestDetails theRequest,
@@ -1587,27 +1598,27 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Subclasses may override this in order to invoke specific operations when
-	 * we're finished handling all the write entries in the transaction bundle
-	 * with a given verb.
-	 */
+     * Subclasses may override this in order to invoke specific operations when
+     * we're finished handling all the write entries in the transaction bundle
+     * with a given verb.
+     */
 	protected void handleVerbChangeInTransactionWriteOperations() {
 		// nothing
 	}
 
 	/**
-	 * Implement to handle post transaction processing
-	 */
+     * Implement to handle post transaction processing
+     */
 	protected void postTransactionProcess(TransactionDetails theTransactionDetails) {
 		// nothing
 	}
 
 	/**
-	 * Check for if a resource id should be matched in a conditional update
-	 * If the FHIR version is older than R4, it follows the old specifications and does not match
-	 * If the resource id has been resolved, then it is an existing resource and does not need to be matched
-	 * If the resource id is local or a placeholder, the id is temporary and should not be matched
-	 */
+     * Check for if a resource id should be matched in a conditional update
+     * If the FHIR version is older than R4, it follows the old specifications and does not match
+     * If the resource id has been resolved, then it is an existing resource and does not need to be matched
+     * If the resource id is local or a placeholder, the id is temporary and should not be matched
+     */
 	private boolean shouldConditionalUpdateMatchId(TransactionDetails theTransactionDetails, IIdType theId) {
 		if (myContext.getVersion().getVersion().isOlderThan(FhirVersionEnum.R4)) {
 			return false;
@@ -1641,9 +1652,9 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * After transaction processing and resolution of indexes and references, we want to validate that the resources that were stored _actually_
-	 * match the conditional URLs that they were brought in on.
-	 */
+     * After transaction processing and resolution of indexes and references, we want to validate that the resources that were stored _actually_
+     * match the conditional URLs that they were brought in on.
+     */
 	private void validateAllInsertsMatchTheirConditionalUrls(
 			Map<IIdType, DaoMethodOutcome> theIdToPersistedOutcome,
 			Map<String, IIdType> conditionalUrlToIdMap,
@@ -1678,12 +1689,12 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Checks for any delete conflicts.
-	 *
-	 * @param theDeleteConflicts  - set of delete conflicts
-	 * @param theDeletedResources - set of deleted resources
-	 * @param theUpdatedResources - list of updated resources
-	 */
+     * Checks for any delete conflicts.
+     *
+     * @param theDeleteConflicts  - set of delete conflicts
+     * @param theDeletedResources - set of deleted resources
+     * @param theUpdatedResources - list of updated resources
+     */
 	private void checkForDeleteConflicts(
 			DeleteConflictList theDeleteConflicts,
 			Set<String> theDeletedResources,
@@ -1733,27 +1744,27 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * This method replaces any placeholder references in the
-	 * source transaction Bundle with their actual targets, then stores the resource contents and indexes
-	 * in the database. This is trickier than you'd think because of a couple of possibilities during the
-	 * save:
-	 * * There may be resources that have not changed (e.g. an update/PUT with a resource body identical
-	 * to what is already in the database)
-	 * * There may be resources with auto-versioned references, meaning we're replacing certain references
-	 * in the resource with a versioned references, referencing the current version at the time of the
-	 * transaction processing
-	 * * There may by auto-versioned references pointing to these unchanged targets
-	 * <p>
-	 * If we're not doing any auto-versioned references, we'll just iterate through all resources in the
-	 * transaction and save them one at a time.
-	 * <p>
-	 * However, if we have any auto-versioned references we do this in 2 passes: First the resources from the
-	 * transaction that don't have any auto-versioned references are stored. We do them first since there's
-	 * a chance they may be a NOP and we'll need to account for their version number not actually changing.
-	 * Then we do a second pass for any resources that have auto-versioned references. These happen in a separate
-	 * pass because it's too complex to try and insert the auto-versioned references and still
-	 * account for NOPs, so we block NOPs in that pass.
-	 */
+     * This method replaces any placeholder references in the
+     * source transaction Bundle with their actual targets, then stores the resource contents and indexes
+     * in the database. This is trickier than you'd think because of a couple of possibilities during the
+     * save:
+     * * There may be resources that have not changed (e.g. an update/PUT with a resource body identical
+     * to what is already in the database)
+     * * There may be resources with auto-versioned references, meaning we're replacing certain references
+     * in the resource with a versioned references, referencing the current version at the time of the
+     * transaction processing
+     * * There may by auto-versioned references pointing to these unchanged targets
+     * <p>
+     * If we're not doing any auto-versioned references, we'll just iterate through all resources in the
+     * transaction and save them one at a time.
+     * <p>
+     * However, if we have any auto-versioned references we do this in 2 passes: First the resources from the
+     * transaction that don't have any auto-versioned references are stored. We do them first since there's
+     * a chance they may be a NOP and we'll need to account for their version number not actually changing.
+     * Then we do a second pass for any resources that have auto-versioned references. These happen in a separate
+     * pass because it's too complex to try and insert the auto-versioned references and still
+     * account for NOPs, so we block NOPs in that pass.
+     */
 	private void resolveReferencesThenSaveAndIndexResources(
 			RequestDetails theRequest,
 			TransactionDetails theTransactionDetails,
@@ -2035,15 +2046,16 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * We should replace the references when
-	 * 1. It is not a reference we should keep the client-supplied version for as configured by `DontStripVersionsFromReferences` or
-	 * 2. It is a reference that has been identified for auto versioning or
-	 * 3. Is a placeholder reference
-	 * @param theReferencesToAutoVersion list of references identified for auto versioning
-	 * @param theReferencesToKeepClientSuppliedVersion list of references that we should not strip the version for
-	 * @param theResourceReference the resource reference
-	 * @return true if we should replace the resource reference, false if we should keep the client provided reference
-	 */
+     * We should replace the references when
+     * 1. It is not a reference we should keep the client-supplied version for as configured by `DontStripVersionsFromReferences` or
+     * 2. It is a reference that has been identified for auto versioning or
+     * 3. Is a placeholder reference
+     *
+     * @param theReferencesToAutoVersion               list of references identified for auto versioning
+     * @param theReferencesToKeepClientSuppliedVersion list of references that we should not strip the version for
+     * @param theResourceReference                     the resource reference
+     * @return true if we should replace the resource reference, false if we should keep the client provided reference
+     */
 	private boolean shouldReplaceResourceReference(
 			Set<IBaseReference> theReferencesToAutoVersion,
 			Set<IBaseReference> theReferencesToKeepClientSuppliedVersion,
@@ -2187,12 +2199,12 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Extracts the transaction url from the entry and verifies it's:
-	 * * not null or blank
-	 * * is a relative url matching the resourceType it is about
-	 * <p>
-	 * Returns the transaction url (or throws an InvalidRequestException if url is not valid)
-	 */
+     * Extracts the transaction url from the entry and verifies it's:
+     * * not null or blank
+     * * is a relative url matching the resourceType it is about
+     * <p>
+     * Returns the transaction url (or throws an InvalidRequestException if url is not valid)
+     */
 	private String extractAndVerifyTransactionUrlForEntry(IBase theEntry, String theVerb) {
 		String url = extractTransactionUrlOrThrowException(theEntry, theVerb);
 
@@ -2206,12 +2218,12 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Returns true if the provided url is a valid entry request.url.
-	 * <p>
-	 * This means:
-	 * a) not an absolute url (does not start with http/https)
-	 * b) starts with either a ResourceType or /ResourceType
-	 */
+     * Returns true if the provided url is a valid entry request.url.
+     * <p>
+     * This means:
+     * a) not an absolute url (does not start with http/https)
+     * b) starts with either a ResourceType or /ResourceType
+     */
 	private boolean isValidResourceTypeUrl(@Nonnull String theUrl) {
 		if (UrlUtil.isAbsolute(theUrl)) {
 			return false;
@@ -2236,9 +2248,9 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Extracts the transaction url from the entry and verifies that it is not null/blank
-	 * and returns it
-	 */
+     * Extracts the transaction url from the entry and verifies that it is not null/blank
+     * and returns it
+     */
 	private String extractTransactionUrlOrThrowException(IBase nextEntry, String verb) {
 		String url = myVersionAdapter.getEntryRequestUrl(nextEntry);
 		if (isBlank(url)) {
@@ -2295,14 +2307,14 @@ public abstract class BaseTransactionProcessor {
 	}
 
 	/**
-	 * Transaction Order, per the spec:
-	 * <p>
-	 * Process any DELETE interactions
-	 * Process any POST interactions
-	 * Process any PUT interactions
-	 * Process any PATCH interactions
-	 * Process any GET interactions
-	 */
+     * Transaction Order, per the spec:
+     * <p>
+     * Process any DELETE interactions
+     * Process any POST interactions
+     * Process any PUT interactions
+     * Process any PATCH interactions
+     * Process any GET interactions
+     */
 	// @formatter:off
 	public class TransactionSorter implements Comparator<IBase> {
