+++ hapi-fhir-storage/src/test/java/ca/uhn/fhir/jpa/util/TransactionSemanticsHeaderTest.java
@@ -14,29 +14,29 @@ class TransactionSemanticsHeaderTest {
 	@ParameterizedTest
 	@ValueSource(strings = {
 		// Same but with different spacing
-		"retryCount=3; minRetryDelay=100; maxRetryDelay=200; finalRetryAsBatch=true",
-		"retryCount =  3 ; minRetryDelay  = 100  ; maxRetryDelay   = 200  ;    finalRetryAsBatch  =   true ",
+		"retryCount=3; minRetryDelay=100; maxRetryDelay=200; tryBatchAsTransactionFirst=true",
+		"retryCount =  3 ; minRetryDelay  = 100  ; maxRetryDelay   = 200  ;    tryBatchAsTransactionFirst  =   true ",
 	})
 	public void testParse(String theInput) {
 		TransactionSemanticsHeader actual = TransactionSemanticsHeader.parse(theInput);
 		assertEquals(3, actual.getRetryCount());
 		assertEquals(100, actual.getMinRetryDelay());
 		assertEquals(200, actual.getMaxRetryDelay());
-		assertTrue(actual.isFinalRetryAsBatch());
+		assertTrue(actual.isTryBatchAsTransactionFirst());
 	}
 
 	@ParameterizedTest
 	@ValueSource(strings = {
-		"retryCount = -1; minRetryDelay = -1; maxRetryDelay = -1; finalRetryAsBatch = -1",
-		"retryCount =; minRetryDelay =; maxRetryDelay =; finalRetryAsBatch =",
-		"retryCount =HELLO; minRetryDelay =HELLO;;;;  ; maxRetryDelay =HELLO; finalRetryAsBatch =HELLO",
+		"retryCount = -1; minRetryDelay = -1; maxRetryDelay = -1; tryBatchAsTransactionFirst = -1",
+		"retryCount =; minRetryDelay =; maxRetryDelay =; tryBatchAsTransactionFirst =",
+		"retryCount =HELLO; minRetryDelay =HELLO;;;;  ; maxRetryDelay =HELLO; tryBatchAsTransactionFirst =HELLO",
 	})
 	public void testParseInvalidValues(String theInput) {
 		TransactionSemanticsHeader actual = TransactionSemanticsHeader.parse(theInput);
 		assertNull(actual.getRetryCount());
 		assertNull(actual.getMinRetryDelay());
 		assertNull(actual.getMaxRetryDelay());
-		assertFalse(actual.isFinalRetryAsBatch());
+		assertFalse(actual.isTryBatchAsTransactionFirst());
 	}
 
 	@Test
@@ -46,11 +46,11 @@ class TransactionSemanticsHeaderTest {
 			.withRetryCount(4)
 			.withMinRetryDelay(100)
 			.withMaxRetryDelay(200)
-			.withFinalRetryAsBatch(true)
+			.withTryBatchAsTransactionFirst(true)
 			.build();
 
 		String actual = header.toHeaderValue();
-		assertEquals("retryCount=4; minRetryDelay=100; maxRetryDelay=200; finalRetryAsBatch=true", actual);
+		assertEquals("retryCount=4; minRetryDelay=100; maxRetryDelay=200; tryBatchAsTransactionFirst=true", actual);
 	}
 
 }
